requirements.txt
# Core dependencies
streamlit>=1.29.0
openai>=1.3.0
biopython>=1.81
tiktoken>=0.5.0
certifi>=2024.2.2

# Logging and utilities
python-dotenv>=1.0.0

# SSL and HTTP
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.9.0

# Time handling
python-dateutil>=2.8.2

# Optional but recommended
tqdm>=4.66.0  # For progress bars 